+++ src/main/java/org/umssdiplo/automationv01/core/customwebdriver/Chrome.java
@@ -2,10 +2,16 @@ package org.umssdiplo.automationv01.core.customwebdriver;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
 
 public class Chrome implements CustomWebDriver {
     @Override
     public WebDriver generateWebDriver() {
-        return new ChromeDriver();
+        System.setProperty("webdriver.chrome.driver","/usr/bin/chromedriver");
+             ChromeOptions options = new ChromeOptions();
+                options.setExperimentalOption("useAutomationExtension", false);
+                options.addArguments("--no-sandbox");
+                options.addArguments("--disable-dev-shm-usage");
+                return new ChromeDriver(options);
     }
 }
