+++ src/main/java/org/umssdiplo/automationv01/core/utils/PropertyAccessor.java
@@ -61,14 +61,6 @@ public class PropertyAccessor {
         return getDataProperty(PASSWORD);
     }
 
-    public String getSmartHouseBaseUrl() {
-        return getDataProperty(SMARTHOUSE_URL);
-    }
-
-    public String getSHUser() { return getDataProperty(SMARTHOUSE_USER); }
-
-    public String getSHPassword() { return getDataProperty(SMARTHOUSE_PASSWORD); }
-
     public String getSHDataProperty(String data) {
         return getDataProperty(data);
     }
