--- src/test/java/org/umssdiplo/automationv01/SHCucumberRunner.java
@@ -1,15 +1,0 @@
-package org.umssdiplo.automationv01;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.testng.AbstractTestNGCucumberTests;
-
-@CucumberOptions(
-        features = {"src/test/resources/smartHouseFeatures/Accident.feature"},
-        format = {"pretty",
-                "html:build/cucumber",
-                "json:build/cucumber/cucumber.json",
-                "junit:build/test-report.xml"},
-        tags = {"@natalynanda"}
-)
-public class SHCucumberRunner extends AbstractTestNGCucumberTests {
-}
